Rename local crypto module to hashlib
Using crypto for the module name is confusing as previously we were using a
crate with the same name, and we want to be clear that this is not a crate we
depend on.
