+++ validator/src/journal/candidate_block.rs
@@ -22,7 +22,7 @@ use cpython::ObjectProtocol;
 use cpython::PyClone;
 use cpython::Python;
 
-use crypto::sha256_digest_strs;
+use hashlib::sha256_digest_strs;
 
 use batch::Batch;
 use transaction::Transaction;
